Isolate post rendering
